clean up onboarding  - create reusable component for email screen  - create reusable component for otp screen  - implement hide keyboard  - implement keyboard avoiding view
