+++ app/(signup)/index.tsx
@@ -1,9 +1,9 @@
 import { PrivyProvider } from '@privy-io/expo';
-import SignupEmail from '@/screens/signup';
+import { EnterEmail } from '@/components';
 
 const SignupTagScreen = () => (
   <PrivyProvider appId={process.env.EXPO_PUBLIC_PRIVY_APP_ID!} clientId={process.env.EXPO_PUBLIC_PRIVY_CLIENT_ID}>
-    <SignupEmail />
+    <EnterEmail isSignup />
   </PrivyProvider>
 );
