+++ components/enter-email/types.ts
@@ -1,0 +1,3 @@
+interface Props {
+  isSignup?: boolean;
+}
