+++ components/index.ts
@@ -10,6 +10,9 @@ import LQDBottomSheet from './bottom-sheet';
 import LQDBarChart from './bar-chart';
 import LQDAssetSelection from './asset-selection';
 import LQDLoadingStep from './loading-step';
+import EnterEmail from './enter-email';
+import VerifyEmail from './verify-email';
+import LQDKeyboardWrapper from './keyboard-wrapper';
 
 export {
   LQDButton,
@@ -23,4 +26,7 @@ export {
   LQDBottomSheet,
   LQDBarChart,
   LQDAssetSelection,
+  EnterEmail,
+  VerifyEmail,
+  LQDKeyboardWrapper,
 };
