+++ components/keyboard-wrapper.tsx
@@ -1,0 +1,11 @@
+import { Keyboard, KeyboardAvoidingView, Platform, TouchableWithoutFeedback } from 'react-native';
+
+const LQDKeyboardWrapper = ({ children }: { children: React.ReactNode }) => (
+  <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
+    <KeyboardAvoidingView behavior={Platform.OS === 'ios' ? 'padding' : undefined} style={{ flex: 1 }}>
+      {children}
+    </KeyboardAvoidingView>
+  </TouchableWithoutFeedback>
+);
+
+export default LQDKeyboardWrapper;
