--- screens/tag/types.ts
@@ -1,1 +1,0 @@
-type TagFormState = 'base' | 'loading' | 'error' | 'success';
